refactor: move all the network-related logic to the hardhat package
