+++ v-next/hardhat/src/internal/network/provider-errors.ts
